error 5: bugprone-implicit-widening-of-multiplication-result

+++ libs/core/iostream/tests/unit/array.cpp
@@ -12,6 +12,7 @@
 #include <hpx/hpx_main.hpp>
 #include <hpx/modules/iostream.hpp>
 #include <hpx/modules/testing.hpp>
+#include <cstddef>
 
 #include "detail/sequence.hpp"
 #include "detail/temp_file.hpp"
@@ -46,7 +47,8 @@ void array_test()
 
     //--------------stream<array_sink>---------------------------------//
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         stream<array_sink<char>> out(&first[0], &first[0] + first.size());
         write_data_in_chars(out);
         ifstream second(
@@ -56,7 +58,8 @@ void array_test()
     }
 
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         stream<array_sink<char>> out(&first[0], &first[0] + first.size());
         write_data_in_chunks(out);
         ifstream second(
@@ -67,7 +70,8 @@ void array_test()
 
     //--------------random access---------------------------------------------//
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         stream<hpx::iostream::array<char>> io(
             &first[0], &first[0] + first.size());
         HPX_TEST_MSG(test_seekable_in_chars(io),
@@ -75,7 +79,8 @@ void array_test()
     }
 
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         stream<hpx::iostream::array<char>> io(
             &first[0], &first[0] + first.size());
         HPX_TEST_MSG(test_seekable_in_chars(io),
